requirements.txt
python-dotenv==0.20.0
binance-connector==1.13.0
